Hice funciones de attack y de damage Defense, para agregar el uso de defense. Tambien a cada funcion le meti los parametros int ataque, int pokemon1, int pokemon2, para que no tuvieran 0, y que funcionara apropiadamente la funcion.

+++ src/main/java/pokemones/Mounstritos.java
@@ -11,11 +11,11 @@ public abstract class Mounstritos {
     protected double HP;
     protected double maxHP;
     protected ListaDE<Movimientos> ataques;
-    protected int defense;
+    protected float defense;
     protected int ataque;
     protected boolean vivo;
 
-    public Mounstritos(String name, double HP,double maxHP, ListaDE<Movimientos> ataques, int defense, int ataque,String sprite,boolean vivo, String [] debilidades) {
+    public Mounstritos(String name, double HP,double maxHP, ListaDE<Movimientos> ataques, float defense, int ataque,String sprite,boolean vivo, String [] debilidades) {
         this.name = name;
         this.HP = maxHP;
         this.maxHP = maxHP;
@@ -49,7 +49,7 @@ public abstract class Mounstritos {
     public double getMaxHP() {
         return maxHP;
     }
-    public int getDefense() {
+    public float getDefense() {
         return defense;
     }
 
@@ -76,7 +76,7 @@ public abstract class Mounstritos {
         this.HP = HP;
     }
 
-    public void setDefense(int defense) {
+    public void setDefense(float defense) {
         this.defense = defense;
     }
