+++ src/main/java/pokemones/tipoAire.java
@@ -7,7 +7,7 @@ public class tipoAire extends Mounstritos {
 
 
 
-    public tipoAire(String name, double HP, double maxHP, ListaDE<Movimientos> ataques, int defense, int ataque, String sprite, boolean vivo) {
+    public tipoAire(String name, double HP, double maxHP, ListaDE<Movimientos> ataques, float defense, int ataque, String sprite, boolean vivo) {
         super(name, HP, maxHP, ataques, defense, ataque, sprite, vivo, new String[]{"Roca","Fuego"});
 
     }
